Create main.py
Kaun Banega Crorepati (simply KBC English: Who Wants to Be a Crorepati (RS. 1,00,00,000/-)) is an Indian Hindi-language television game show. It is the official Hindi adaptation of the Who Wants to Be a Millionaire? 

In this Python Practice Program. Which is based on KBC's Season 1.

Has 15 Questions - Winning Price rupees 1Cr.

There are 3 stages. A, B, & C

A. This is the beginning of the game Containing a Total 5 Questions where if the player Locks the wrong answer He will go home with nothing. If the Player Completes this stage he will secure Rs. 10,000/- which he will carry even if he loses in 2nd and 3rd Stages.

B. This stage Contains a Total of 5 Questions too, where if the player Locks the wrong answer He will go home with Rs. 10,000\- If the Player Completes this stage he will secure Rs. 3,20,000/- which he will carry even if he loses in the 3rd Stages.

C. The Final Stage, The most difficult and most expensive, This stage Contains a Total of 5 Questions too, where if the player Locks the wrong answer He will go home with Rs. 3,20,000\- which he secured in the last stage, If the Player Completes this stage he will Complete this game with winning price Rs. 1,00,00,000/- which he will carry Home if he Wins.

**This is a Practice project, No price is Real**

Payment Method will be Pixels on the Screen :)